major refactor: project for assemblt reflection, msBuild task to copy its dll to consoleapp
