+++ AssemblyReflection/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Autofac" version="4.8.1" targetFramework="net461" />
+</packages>
