+++ Clairvoyant.sln
@@ -5,6 +5,10 @@ VisualStudioVersion = 15.0.28010.2048
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Core", "Core\Core.csproj", "{EF10C052-5948-4AE9-81FD-72AD7BD610A2}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AssemblyReflection", "AssemblyReflection\AssemblyReflection.csproj", "{1EFA8609-976D-4896-A711-672760FD8931}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DumpConsoleApp", "DumpConsoleApp\DumpConsoleApp.csproj", "{8629D610-3CD7-475C-BDC7-01EE1740D589}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -15,6 +19,14 @@ Global
 		{EF10C052-5948-4AE9-81FD-72AD7BD610A2}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{EF10C052-5948-4AE9-81FD-72AD7BD610A2}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{EF10C052-5948-4AE9-81FD-72AD7BD610A2}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1EFA8609-976D-4896-A711-672760FD8931}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1EFA8609-976D-4896-A711-672760FD8931}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1EFA8609-976D-4896-A711-672760FD8931}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1EFA8609-976D-4896-A711-672760FD8931}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8629D610-3CD7-475C-BDC7-01EE1740D589}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8629D610-3CD7-475C-BDC7-01EE1740D589}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8629D610-3CD7-475C-BDC7-01EE1740D589}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8629D610-3CD7-475C-BDC7-01EE1740D589}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
