+++ Core/Components/IDataSource.cs
@@ -1,0 +1,9 @@
+﻿using Core.Model;
+
+namespace Core.Components
+{
+    public interface IDataSource
+    {
+        AssemblyMetadata GetAssemblyMetadata(string assemblyFile);
+    }
+}
