+++ Core/Components/ILogger.cs
@@ -1,0 +1,7 @@
+﻿namespace Core.Components
+{
+    public interface ILogger
+    {
+        void Info(string message);
+    }
+}
