+++ Core/Constants/TypeKind.cs
@@ -1,0 +1,10 @@
+﻿namespace Core.Constants
+{
+    public enum TypeKind
+    {
+        EnumType,
+        StructType,
+        InterfaceType,
+        ClassType
+    }
+}
