+++ Core/Core.csproj
@@ -39,6 +39,9 @@
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <ItemGroup>
+    <Reference Include="Autofac, Version=4.8.1.0, Culture=neutral, PublicKeyToken=17863af14b0044da, processorArchitecture=MSIL">
+      <HintPath>..\packages\Autofac.4.8.1\lib\net45\Autofac.dll</HintPath>
+    </Reference>
     <Reference Include="System" />
     <Reference Include="System.Core" />
     <Reference Include="System.Xml.Linq" />
@@ -49,13 +52,17 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Components\IDataSource.cs" />
+    <Compile Include="Components\ILogger.cs" />
+    <Compile Include="Model\BaseRegisterModule.cs" />
     <Compile Include="Constants\Accessibility.cs" />
     <Compile Include="Constants\IsAbstract.cs" />
     <Compile Include="Constants\IsSealed.cs" />
     <Compile Include="Constants\IsStatic.cs" />
     <Compile Include="Constants\IsVirtual.cs" />
-    <Compile Include="ExtensionMethods\MethodBaseExtensions.cs" />
-    <Compile Include="ExtensionMethods\TypeExtensions.cs" />
+    <Compile Include="Constants\TypeKind.cs" />
+    <Compile Include="ContainerProvider.cs" />
+    <Compile Include="GlobalSuppressions.cs" />
     <Compile Include="Model\AssemblyMetadata.cs" />
     <Compile Include="Model\MethodMetadata.cs" />
     <Compile Include="Model\NamespaceMetadata.cs" />
@@ -79,7 +86,9 @@
     <Analyzer Include="..\packages\Text.Analyzers.2.6.2\analyzers\dotnet\cs\Text.Analyzers.dll" />
     <Analyzer Include="..\packages\Text.Analyzers.2.6.2\analyzers\dotnet\cs\Text.CSharp.Analyzers.dll" />
   </ItemGroup>
-  <ItemGroup />
+  <ItemGroup>
+    <Folder Include="Exceptions\" />
+  </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Target Name="EnsureNuGetPackageBuildImports" BeforeTargets="PrepareForBuild">
     <PropertyGroup>
