+++ Core/Model/AssemblyMetadata.cs
@@ -1,24 +1,17 @@
-﻿using Core.ExtensionMethods;
-using System;
+﻿using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Reflection;
 
 namespace Core.Model
 {
-    internal class AssemblyMetadata
+    public class AssemblyMetadata
     {
-        internal AssemblyMetadata(Assembly assembly)
-        {
-            _name = assembly.ManifestModule.Name;
-            _namespaces = from Type type in assembly.GetTypes()
-                          where type.IsVisible()
-                          group type by type.GetNamespace() into namespaceGroup
-                          orderby namespaceGroup.Key
-                          select new NamespaceMetadata(namespaceGroup.Key, namespaceGroup);
-        }
+        public string Name { get; }
+        public IEnumerable<NamespaceMetadata> Namespaces { get; }
 
-        private string _name;
-        private IEnumerable<NamespaceMetadata> _namespaces;
+        public AssemblyMetadata(string name, IEnumerable<NamespaceMetadata> namespaces)
+        {
+            Name = name ?? throw new ArgumentNullException(nameof(name));
+            Namespaces = namespaces ?? throw new ArgumentNullException(nameof(namespaces));
+        }
     }
 }
