+++ Core/Model/BaseRegisterModule.cs
@@ -1,0 +1,9 @@
+﻿using Autofac;
+
+namespace Core.Model
+{
+    public abstract class BaseRegisterModule : Module
+    {
+        protected abstract override void Load(ContainerBuilder builder);
+    }
+}
