+++ Core/Model/PropertyMetadata.cs
@@ -1,26 +1,14 @@
-﻿using Core.ExtensionMethods;
-using System.Collections.Generic;
-using System.Linq;
-using System.Reflection;
-
-namespace Core.Model
+﻿namespace Core.Model
 {
-    internal class PropertyMetadata
+    public class PropertyMetadata
     {
-        private readonly string _name;
-        private readonly TypeMetadata _typeMetadata;
+        public string Name { get; }
+        public TypeMetadata Metadata { get; }
 
         private PropertyMetadata(string propertyName, TypeMetadata propertyType)
         {
-          _name = propertyName;
-          _typeMetadata = propertyType;
-        }
-
-        internal static IEnumerable<PropertyMetadata> EmitProperties(IEnumerable<PropertyInfo> props)
-        {
-          return from prop in props
-                 where prop.GetGetMethod().IsVisible() || prop.GetSetMethod().IsVisible()
-                 select new PropertyMetadata(prop.Name, TypeMetadata.EmitReference(prop.PropertyType));
+          Name = propertyName;
+          Metadata = propertyType;
         }
     }
 }
