+++ Core/packages.config
@@ -1,5 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
+  <package id="Autofac" version="4.8.1" targetFramework="net461" />
   <package id="Microsoft.CodeAnalysis.FxCopAnalyzers" version="2.6.2" targetFramework="net461" developmentDependency="true" />
   <package id="Microsoft.CodeQuality.Analyzers" version="2.6.2" targetFramework="net461" developmentDependency="true" />
   <package id="Microsoft.NetCore.Analyzers" version="2.6.2" targetFramework="net461" developmentDependency="true" />
